Reorder module layout
