+++ singer/logging.py
@@ -1,0 +1,11 @@
+import logging
+import logging.config
+import os
+
+
+def get_logger():
+    """Return a Logger instance appropriate for using in a Tap or a Target."""
+    this_dir, _ = os.path.split(__file__)
+    path = os.path.join(this_dir, 'logging.conf')
+    logging.config.fileConfig(path)
+    return logging.getLogger('root')
